+++ Tests/HummingbirdFoundationTests/FilesTests.swift
@@ -53,6 +53,27 @@ class HummingbirdFilesTests: XCTestCase {
         }
     }
 
+    func testRead() async throws {
+        let app = HBApplicationBuilder(requestContext: HBTestRouterContext.self)
+        app.router.get("test.jpg") { _, context -> HBResponse in
+            let fileIO = HBFileIO(threadPool: context.applicationContext.threadPool)
+            let body = try await fileIO.loadFile(path: "test.jpg", context: context, logger: context.logger)
+            return .init(status: .ok, headers: [:], body: body)
+        }
+        let buffer = self.randomBuffer(size: 320_003)
+        let data = Data(buffer: buffer)
+        let fileURL = URL(fileURLWithPath: "test.jpg")
+        XCTAssertNoThrow(try data.write(to: fileURL))
+        defer { XCTAssertNoThrow(try FileManager.default.removeItem(at: fileURL)) }
+
+        try await app.buildAndTest(.router) { client in
+            try await client.XCTExecute(uri: "/test.jpg", method: .GET) { response in
+                XCTAssertEqual(response.body, buffer)
+            }
+        }
+    }
+
+
     func testReadLargeFile() async throws {
         let router = HBRouterBuilder(context: HBTestRouterContext.self)
         router.middlewares.add(HBFileMiddleware("."))
