+++ Tests/HummingbirdTests/PersistTests.swift
@@ -12,7 +12,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-@testable import Hummingbird
+import Hummingbird
 import HummingbirdXCT
 import XCTest
 
@@ -26,14 +26,15 @@ final class PersistTests: XCTestCase {
         router.put("/persist/:tag") { request, context -> HTTPResponseStatus in
             guard let tag = context.parameters.get("tag") else { throw HBHTTPError(.badRequest) }
             guard let buffer = request.body.buffer else { throw HBHTTPError(.badRequest) }
+            let tag = try context.parameters.require("tag")
             try await persist.set(key: tag, value: String(buffer: buffer), request: request)
             return .ok
         }
         router.put("/persist/:tag/:time") { request, context -> HTTPResponseStatus in
             guard let time = context.parameters.get("time", as: Int.self) else { throw HBHTTPError(.badRequest) }
-            guard let tag = context.parameters.get("tag") else { throw HBHTTPError(.badRequest) }
             guard let buffer = request.body.buffer else { throw HBHTTPError(.badRequest) }
-            try await persist.set(key: tag, value: String(buffer: buffer), expires: .seconds(numericCast(time)), request: request).get()
+            let tag = try context.parameters.require("tag")
+            try await persist.set(key: tag, value: String(buffer: buffer), expires: .seconds(numericCast(time)), request: request)
             return .ok
         }
         router.get("/persist/:tag") { request, context -> String? in
@@ -67,7 +68,8 @@ final class PersistTests: XCTestCase {
         router.put("/create/:tag") { request, context -> HTTPResponseStatus in
             guard let tag = context.parameters.get("tag") else { throw HBHTTPError(.badRequest) }
             guard let buffer = request.body.buffer else { throw HBHTTPError(.badRequest) }
-            try await persist.create(key: tag, value: String(buffer: buffer), request: request).get()
+            let tag = try context.parameters.require("tag")
+            try await persist.create(key: tag, value: String(buffer: buffer), request: request)
             return .ok
         }
         let app = HBApplication(responder: router.buildResponder())
@@ -86,11 +88,12 @@ final class PersistTests: XCTestCase {
         router.put("/create/:tag") { request, context -> HTTPResponseStatus in
             guard let tag = context.parameters.get("tag") else { throw HBHTTPError(.badRequest) }
             guard let buffer = request.body.buffer else { throw HBHTTPError(.badRequest) }
-            try await persist.create(key: tag, value: String(buffer: buffer), request: request)
-                .flatMapErrorThrowing { error in
-                    if let error = error as? HBPersistError, error == .duplicate { throw HBHTTPError(.conflict) }
-                    throw error
-                }.get()
+            let tag = try context.parameters.require("tag")
+            do {
+                try await persist.create(key: tag, value: String(buffer: buffer), request: request)
+            } catch let error as HBPersistError where error == .duplicate {
+                throw HBHTTPError(.conflict)
+            }
             return .ok
         }
         let app = HBApplication(responder: router.buildResponder())
@@ -144,6 +147,10 @@ final class PersistTests: XCTestCase {
     }
 
     func testCodable() async throws {
+        #if os(macOS)
+        // disable macOS tests in CI. GH Actions are currently running this when they shouldn't
+        guard HBEnvironment().get("CI") != "true" else { throw XCTSkip() }
+        #endif
         struct TestCodable: Codable {
             let buffer: String
         }
@@ -151,12 +158,13 @@ final class PersistTests: XCTestCase {
         router.put("/codable/:tag") { request, context -> HTTPResponseStatus in
             guard let tag = context.parameters.get("tag") else { throw HBHTTPError(.badRequest) }
             guard let buffer = request.body.buffer else { throw HBHTTPError(.badRequest) }
-            try await persist.set(key: tag, value: TestCodable(buffer: String(buffer: buffer)), request: request).get()
+            try await persist.set(key: tag, value: TestCodable(buffer: String(buffer: buffer)), request: request)
             return .ok
         }
         router.get("/codable/:tag") { request, context -> String? in
             guard let tag = context.parameters.get("tag") else { throw HBHTTPError(.badRequest) }
-            return try await persist.get(key: tag, as: TestCodable.self, request: request).get()?.buffer
+            let value = try await persist.get(key: tag, as: TestCodable.self, request: request)
+            return value?.buffer
         }
         let app = HBApplication(responder: router.buildResponder())
